Switch graph accumulator to use new support calculator (#1013)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1013

We need to know the support -- the set of possible values of a stochastic expression -- to handle compiling models of the form

    x = some_rv(some_finite_rv())

where some_finite_rv has small, finite support. It needs to be small and finite because we're going to actually call some_rv with each possible value.

For our purposes we define small as <= 1000.  (We should add a mechanism to tweak this parameter later.)

Previously we had each node compute its own support, but (1) this is not a concern of the node, and (2) we need extra mechanisms to also compute approximate support size, and (3) the implementation was recursive, which crashes if the model has a path that exceeds python's recursion limit.

We now have a module using the same nonrecurisve, mutation-aware algorithm that the type checker uses. In this diff I switch over the graph accumulator to use the new mechanism.  In the next diff I will delete the now-unused support instance methods.

Reviewed By: wtaha

Differential Revision: D30738627

fbshipit-source-id: 5e745f272d59844acd44e1ba2e36b13160dee42c