Add more clear precondition verification to requirements checking pass
Summary:
tldr: the requirements fixer has complex preconditions that are easy to violate. This diff makes it easier on compiler developers to figure out what they've forgotten to do when the requirements fixer fails, by (1) failing early, and (2) producing a helpful error message.

excessive details:

The purpose of the problem fixing passes is to transform the graph into one where every ancestor of a query or observation is (1) supported by BMG, and (2) each edge in the graph that is on a path to a query or observation meets the requirements of the BMG type system.  That is to say: when the problem fixing passes are done, we should either have produced an error, or we can produce a BMG graph without causing BMG to throw an exception.

The requirements fixer checks to see if every such edge meets BMG's requirements; if an edge does not meet a requirement then the requirements checker either produces an error, or it mutates the graph to remove the offending edge. (For example, by inserting a TO_REAL node and rerouting the outgoing edge through it.)

The requirements fixer has many preconditions. It assumes:

* Every type requirement imposed on an edge is for an actual valid BMG type. For example, type analysis might determine that a node is a 3-dimensional tensor, but BMG only supports 2-dimensional matrices. Since the whole point of the requirement fixer is to make the graph compliant with the BMG type system, it makes no sense to ask it to find a way to make an edge point from a node with an unsupported type. If we ever ask to impose such a requirement, there is probably a bug in the code which computes requirements.

* Every node which has a requirement on its outgoing edge is a node that *has* an outgoing edge! We should never be imposing a requirement on a query, observation or factor, for instance, because they never have outgoing edges to begin with. If the compiler is ever trying to check outgoing edge requirements on such a node, then something has gone very wrong with the graph topology.

* Every node (except constants, which are fixed by the requirements checker itself) that has an outgoing edge being checked for validity is already a node supported by BMG. Again, the purpose of the requirements fixer is to make the edges right; if we have a node in the graph that is not even a valid BMG node then it makes no sense to check its edges; the graph cannot possibly be right. The requirements checker expects that the unsupported node fixer has already run, and has either produced an error (in which case the requirements checker should not run at all) or successfully removed all the unsupported nodes.

* Every node (except constants) can have a type assigned to it by the lattice typer. Even if every outgoing edge has a requirement of "any", we still want to maintain the invariant that all BMG nodes in the graph have a BMG node type associated with them. If they don't, that's evidence of a bug in the lattice typer.

Reviewed By: wtaha

Differential Revision: D31175997

fbshipit-source-id: b359fe60c41c24bb8708a8a630ea4b6a08dd564f