Remove now-unused size property from node class
Summary:
We need the tensor size of a node during graph rewriting, particularly when devectorizing models.

I wish to move computation of the tensor size of a node out of the node classes; this is not their concern, and the implementation needs to be nonrecursive. All this functionality is now moved to sizer.py and I can delete the dead implementations on the node classes.

Reviewed By: wtaha

Differential Revision: D30811484

fbshipit-source-id: 1db9d9040e72b5c7e5883f41b243894504caadc2