Remove now-unused size property from node class (#1019)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1019

We need the tensor size of a node during graph rewriting, particularly when devectorizing models.

I wish to move computation of the tensor size of a node out of the node classes; this is not their concern, and the implementation needs to be nonrecursive. All this functionality is now moved to sizer.py and I can delete the dead implementations on the node classes.

Reviewed By: wtaha

Differential Revision: D30811484

fbshipit-source-id: 447af5747afa68b329059d580304a89f76ec16a4