Factor sequential stepper functionality out of MH.
Summary:
MH was responsible for taking a vector of single-site stepping methods, building the appropriate single-site steppers, keeping them in a vector, and iterating over that vector in order to generate a sample. We now that a `SequentialSingleSiteStepper` class that does all that. From now on, MH just has a single stepper which is invoked when it's time to generate a sample.

Currently, there is still an assumption that MH's stepper is a `SequentialSingleSiteStepper`, and MH is still responsible for taking the vector of single-site stepping methods and passing them on to the constructor of `SequentialSingleSiteStepper`. This will be lifted in the next step, when MH simply received an already constructed stepper.

Differential Revision: D31132804

fbshipit-source-id: 721269527a61d5634fe46d2f30dc5cfeff27bf6f