Replace single-site steppers in MH by regular steppers.
Summary:
MH was currently finding the appropriate single-site stepper for each variable, storing that association, and applying these single-site steppers to their variables at each step.

To make MH more flexible, we want it to use steppers (which are not necessarily single-site) instead of only single-site steppers. So here we introduce two new classes: the abstract class `Stepper`, which has a method `step()` without arguments (unlike `SingleSiteStepper` which requires a variable node to be provided), and a class `BoundStepper` which binds a single-site stepper to a variable node, this creating a `Stepper` out of that pair. MH then creates a list of `Stepper`s to apply at each step.

In the next changes, we will abstract the creation of those steppers away from MH.

Differential Revision: D30829721

fbshipit-source-id: 7add73f2bf5c351304c3e516eb78d8b482cdd24d