Refactor devectorizer prototype to generalize it
Summary:
My devectorizer prototype only supports devectorizing samples on Bernoulli distributions. We need to do much better than this but I'd rather not cut-n-paste a lot of similar code.

This refactoring organizes the code better and will make it much easier to generalize the devectorizing operation to other distributions and operators.

The two basic ideas here are:

* we need to get indexes into the inputs to the distribution.  We now put those into a list of index nodes.

* we need to construct INDEX --> DISTRIBUTION --> SAMPLE nodes for each, where the distribution is unvectorized. We look up the factory method for each distribution we rewrite from a table.

This refactoring will make it much easier to extend the devectorizing algorithm to distributions with two or more inputs, such as normal. That will happen in upcoming diffs.

Reviewed By: wtaha

Differential Revision: D30880949

fbshipit-source-id: e30a7199a8dbb5d35efe55f8c131279d2540e052