Factor sequential stepper functionality out of MH. (#1023)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1023

MH was responsible for taking a vector of single-site stepping methods, building the appropriate single-site steppers, keeping them in a vector, and iterating over that vector in order to generate a sample. We now that a `SequentialSingleSiteStepper` class that does all that. From now on, MH just has a single stepper which is invoked when it's time to generate a sample.

Currently, there is still an assumption that MH's stepper is a `SequentialSingleSiteStepper`, and MH is still responsible for taking the vector of single-site stepping methods and passing them on to the constructor of `SequentialSingleSiteStepper`. This will be lifted in the next diff, when MH simply receives an already constructed stepper.

Reviewed By: yucenli

Differential Revision: D31132804

fbshipit-source-id: f53f3546a1da79f5bd858e3780e9fd76b4a0219a