Adding graph structure metadata to the new world (#1017)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1017

In multi-site inference, we often need to compute the joint prob on a small subgraph (since most of the graph is fixed, this is supposed to be more efficient than recomputing the joint prob of the entire graph per proposal).

This diff adds the parent-children metadata similar to what we had in the [single site world](https://www.internalfb.com/code/fbsource/fbcode/beanmachine/beanmachine/ppl/world/world.py). Note that in this diff, an edge is only added during initialization, so it won't capture change in the structure (as this requires re-evaluating the model for every change in the graph -- which will be added in D31039733).

-----

I also decided to make the world immutable by remove the `__setitem__` and `__delitem__` method from world. I kept `set_transform` in this diff because NUTS and HMC still depends on it, but it will also be removed in D31039733.

`__delitem__` doesn't really make sense in this context, because even if we remove a node from a world, the implicit dependency still exist.

`__setitem__` and `set_transform` will be replaced by some corresponding `replace` method, which will return *a new world* with updated graph structure instead of mutating the current one.

Basically, each world represent a snapshot of the current state. With each proposal, we will be getting a new snapshot (instead of keep updating the old one).

-----

This diff is part of the early draft of the multisite inference. The API that are presented here are only temporary and could change a lot with the upcoming diffs. Feel free to make suggestions :)

Reviewed By: jpchen, neerajprad

Differential Revision: D30984120

fbshipit-source-id: 739b62bf014dff991b1682d1b2bc37ddd4a0eece