Generalize the devectorizer (#1029)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1029

In this diff I finally take a big step towards generalizing the devectorizing algorithm I've been prototyping in this stack.

The general algorithm here relies upon a commonality amongst many operators and distributions in our system:

* the output size results from broadcasting all the inputs
* logically, every input is broadcast to that size

The actual algorithm is a little tricky and so I have left comments describing its action on a test case. The basic idea is that we create tensors of integers representing offsets into a list of nodes. By broadcasting those indexing tensors, we can make torch do the work of figuring out the broadcasting semantics for us.

Thus far I have only tried out this new algorithm on normals and Bernoullis, but it works quite nicely. I will add more distributions and operators in upcoming diffs.

The output graph is a valid BMG graph but is highly non-optimal. In an upcoming diff I will add an optimizing pass that eliminates the unnecessary temporary matrix.

Reviewed By: wtaha

Differential Revision: D30883037

fbshipit-source-id: 0812460635298709896f76bf5c074996c5ea64f1