Optimize away multiple index operations
Summary:
Previously, the graph optimizer would look for situations like `tensor([X, Y, Z])[1]` and replace it with just `Y`. However, this did not work with two-dimensional tensors.

We now look for situations like `tensor([[A, B], [C, D]])[0][0]` and replace it with just `A`.

If you take a look at the before and after for the devectorizer test case you can see how this works. BEFORE:

{F664174819}

We have six samples from Bernoullis, put them into a matrix M and then read right back out M[0][0], M[0][1], ... and so on.  Why have the matrix at all? AFTER:

{F664174932}

Much simpler.

But the big win is with our CLARA test case, which uses this pattern of "put a bunch of samples into a matrix and then index into it with constants" all the time. As you can see in the diff, this reduces the number of nodes from 201 to 168 and reduces the number of edges from 345 to 280. I have not yet calculated by how much it reduces the total descendant count of each sample but probably by rather a lot, and our perf is gated on this quantity.

Reviewed By: wtaha

Differential Revision: D30965161

fbshipit-source-id: daf18a42aa011490f7dd466953a1f9f8337da06b