Pulling the matrix_scale fixer forward in the list of standard fixers
Summary: Earlier in this stack we implement a matrix_scale fixer that works by introduce a matrix_scale node when it sees a *binary* multiplication with one scalar and one matrix argument. As such, it assumes that multiplications are binary, which currently is only reasonable to assume before the MultiaryMulitplication fixer is applied. However, it is concievable that in the future we can find ways to do this transformation with multiary multiplications.

Reviewed By: ericlippert

Differential Revision: D31345003

fbshipit-source-id: 7c49c49f1c7015a489d97561db8bb240b612c6cc