Forward mode AD for MATRIX_SCALE
Summary: Previously in this diff stack, which is implemented the MATRIX_SCALE operation, I had implemented backward AD but not forward. This diff provides a basic implementation along with a basic test. The implementation includes a TODO that I would like to get back to once the e2e functionality for MATRIX_SCALE is realized.

Reviewed By: yucenli

Differential Revision: D31224451

fbshipit-source-id: 206147bb376b163ff6d0bf6f96ddaadc2e823eeb