Add fixer for matrix_scale
Summary:
This diff adds a fixer that recognizes (binary) multiplications where one argument is a matrix and the other is a scalar, and generates a matrix_scale node for that case.

Two tests have been added. The first starts off with the following graph representation

{F667152139}

which then gets transformed to the following representation as a result of the new fixer:

{F667152175}

We also have a second test case that starts with the following representation:

{F667152415}

Which produces the following graph:

{F667154485}

Reviewed By: ericlippert

Differential Revision: D31124345

fbshipit-source-id: b1a9a2f36ff892b4a15083e12bcb33bcb40c25fd