Copying and renmaing MATRIX_MULTIPLE operations to prepare for implementing MATRIX_SCALE
Summary: This diff is the second in a sequence intended to implement matrix scaling (multiplying a matrix by a scalar). In this one we simply make a copy and rename of the codes that we plan to adapt. The idea is that this will make future diffs easier to read, as they will only contain actual modifications and not copying or renmaing (which adds noise when we are looking to see just the changes).

Reviewed By: rodrigodesalvobraz

Differential Revision: D30138734

fbshipit-source-id: c7323559405d79329feb3eba8b3c4c7cf137ea3e