Devectorize addition operator (#1036)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1036

In the previous diffs in this stack we successfully devectorize sample operations on most distributions. In this and the next few diffs we extend this work to operators, starting with addition. If we accumulate this graph:

{F664628900}

then we need to devectorize both the beta sample into two samples, and the addition into four additions:

{F664628934}

Reviewed By: wtaha

Differential Revision: D31007062

fbshipit-source-id: 4ed5b46169150b5185e1196a9a0132557fe1d5f1