Free form initialize_fn (#1020)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1020

This diff introduces an `initialize_fn` arg that replaces the current  `initialize_from_prior` flag. The default behavior is to initialize to Uniform(-2, 2) in the unconstrained space, but users are free to define their own initialization scheme.

The change in this diff is completely optional and doesn't block the rest of diffs in the stack. It just something that I consider as a small improvement to our framework :). Feel free to let me know if you think it's better not to make the change at all.

Note that the change in this diff only affects inferences defined under `global_inference` directory. I can also make the change on the rest of the libraries to keep the API consistent.

Reviewed By: jpchen, neerajprad

Differential Revision: D31010510

fbshipit-source-id: 7635b5a308b8cc1c4a969593149bfcb77afd0519