Fixes to flakey test
Summary:
Recently, TestX and CircleCI have been reporting these tests as flakey. There seems to be a couple of things going on:
1. Some tests sometimes fails producing a value outside the delta used in different tests. The failing delta is about .56.
2. One of the tests times out

This diff relaxes the delta to deal with the first issue. This is done by seeting the detla in the test to 0.75. Interestingly, while the failing delta is different from what it is on devserver tests, it seems to be exactly the same value. So, it seems that there is a specific context in the TestX environment that leads to a different seed being set. Also, in the particular situation of the tests that we do here on mean min and max, it seems that essentially any delta less than 1 would still be providing useful information (essentially, that the delta is within some interval between -2 and 2.

Reviewed By: jpchen

Differential Revision: D31481395

fbshipit-source-id: 489910b2a9f67ada99316da735316ce2ee7566cb