Devectorize more operators (#1037)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1037

In this diff I devectorize more of the operators; we now devectorize `+ - * ** / exp log phi logistic` operators. The remaining operators I'll do in a later diff.

As in previous diffs, I've used a table-driven approach to cut down on duplicated code.

While testing I realized that I had forgotten an important case in previous diffs: operations where one operand is a scalar constant. Those constants have have size `[]`, not size `[1]` and my code did not take that into account. It now does and we test this scenario.

UPDATE: We discovered that the original version of this diff interacts poorly with the matrix scale rewriter, for two reasons.

The first, straightforward reason is: there is no need to devectorize multiplication of a matrix by a scalar because we can now generate a MATRIX_SCALE operator for that.  We simply disable the devectorizer if we have a multiplication of a matrix by a scalar.

This then led to a second problem: we have a chicken-and-egg ordering situation in the rewriters. This is fundamentally an architectural problem that I will fix in a later series of diffs; the principled solution is to run every rewriter until they all collectively reach a fixpoint, and *then* check for errors. (Which is what we do for syntactic rewrites.)

Rather than solve that architectural problem now, I've temporarily merged the matrix scale rewriter and the unsupported node rewriter.

See the comments in fix_unsupported.py for more details.

Reviewed By: wtaha

Differential Revision: D31034257

fbshipit-source-id: dcc99b9507c62a3018c8b49f3d33d773d9494de9