Tutorial on Robust Linear Regression as a Beanstalk test case
Summary: This diff is part of a statck that introduces tutorials as Beanstalk test cases and (later) mitigations to any issues encountered in running these tests through Beanstalk. Currently, in this diff we encountered a situation where the original tutorial (using Bean Machine) seems to generate a runtime error. This diff being shared to help figure fix this problem so that we can proceed to testing Beanstalk.

Reviewed By: ericlippert

Differential Revision: D28121007

fbshipit-source-id: 20eed571d313d8565d0be52e6bf2482440ee013c