New error in the Robust Linear Regression tutorial
Summary: Previous diffs identified and addressed one problem with the Robust Linear Regression tutorial, namely, the need for having a BMG implementation of the half_normal distribution. Once this was addressed, another problem surfaced, and is captured the the error message recorded as a comment in this diff (in place of the previous error message).

Differential Revision: D29322523

fbshipit-source-id: 809c235c33dedd896036b756eb4bf916862c1a1f