Moving CLARA models to a local `facebook` directory
Summary: Currently we have some CLARA models as end to end tests for the compiler. It is very useful to have such realistic models as unit tests that we run regularly to make sure that nothing breaks these important models. At the same time, we are nearing OSS release, and it is not our intent to publish these models at this time. In this diff we exclude CLARA models from the upcoming OSS release by creating a directory called `facebook` and moving the clara*_test.py files there. Directories called `facebook` are removed during the automated open source release process. There are already a couple of other facebook directories, but we want to keep these files local. This way they will continue to be easy to find for compiler developers, and to be part of our internal automated test suite. As indicated by the result of running the Test Plan below, the tests still run internally.

Reviewed By: kshah1997

Differential Revision: D31504888

fbshipit-source-id: a04c093bf7486e4d150696328a8997ec8c2197be