New error in the Robust Linear Regression tutorial (#1045)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1045

Previous diffs identified and addressed one problem with the Robust Linear Regression tutorial, namely, the need for having a BMG implementation of the half_normal distribution. Once this was addressed, another problem surfaced, and is captured the the error message recorded as a comment in this diff (in place of the previous error message).

Reviewed By: ericlippert

Differential Revision: D29322523

fbshipit-source-id: f2e3c60109fc870905e3732b49326ceb5126d17f