Start refactoring assignment rewriting rules (#1047)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1047

The assignment rewriting rules will need to be refactored to be more general, so that they can handle the augmented assignment operators (`+=` and so on.)

I'm going to start by refactoring the left-hand assignment rewriter pattern factories to reduce the amount of duplicated code, and thereby make each pattern factory contain only the information necessary to build the rule. We can then change the more general rule builder to handle any kind of assignment in a later diff.

Right now I'm just doing a single pattern factory to see how it looks; I'll then extend this to the rest of the left-hand pattern factories and the right-hand pattern factories in later diffs.

Reviewed By: wtaha

Differential Revision: D31223315

fbshipit-source-id: d66049a21878926818bf27b2d502d73d67b91fdb