Caching distribution object in world (#1048)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1048

This diff was split from D31259985 (PR#1033) to make the reviewing process a bit easier. In short, what this diff does is to store the distribution objects when we update the values in world, so that we don't need to re-run the model when we need to evaluate the log prob later on.

Since the value of a node and its distribution uniquely determine the value of the log prob, I also make it a cached property.

To ensure that the gradient is still being propagated through the right graph, in HMC proposer, we have to explicitly replace the variables in the world with ones that has grad enabled. This might look a bit redundant, but in terms of computation cost, this is not worse than previous diff (where computing log prob itself requires re-executing the model). The computation cost should be further reduced in D31259985 where HMC proposer only needs to deal with raw dictionaries.

Reviewed By: neerajprad

Differential Revision: D31493782

fbshipit-source-id: 3e7f5072f013f4c6a7ddf713ce225e0be96a470e