Start implementing rewrites of augmented assignments
Summary:
The Beanstalk compiler starts by translating model source code into a simpler subset of Python that is more easily instrumented; we do not yet transform augmented assignments (`+=` and so on) into a simpler form.  This diff begins by ensuring that if the left side of an augmented assignment is an identifier then so is the right side.

In upcoming diffs we will ensure that the left side is always an identifier; that in combination with this rule will achieve the desired final state: every augmented assignment has identifiers on both sides.

Reviewed By: wtaha

Differential Revision: D31283089

fbshipit-source-id: 1a8ff1825cef8839d902c17403ec7706c0f69b47