Continue implementing augmented assignment syntactic rewriter
Summary:
In this diff I continue the work started in the previous; here we implement two more rules for simplifying augmented assignments. They are:

* `complex.attr += any` is rewritten to `a0 = complex; a0.attr += any`
* `a0.attr += any` is rewritten to `a1 = a0.attr ; a1 += any; a0.attr = a1`

In the previous diff we wrote a rule where `a1 += any` is rewritten to `a2 = any; a1 += a2` and so we now have almost all the rules we need to ensure that every augmented assignment only has identifiers on both sides.

We do not yet implement rules for `a[b] += c` and the like; those will come in a later diff.

Reviewed By: wtaha

Differential Revision: D31293128

fbshipit-source-id: fb2eda35ab1bfecdbf7fde36ae70b27ca69c7f2b