Start implementing support for passing in number of chains from Python side to BMG
Summary: Currently the BMGInference.infer() method does not support an argument for specifying the number of chains. In this diff we take the first step towards providing this functionality. First, we change the C++ side method that we use for inference. There are two such methods, and one simply has takes fewer parameters (and returns a result type that assumes there is only one chain). While we are at it we add a line about performance reporting that was in one C++ infer method to the other C++ infer method. We also add a small unit test to assist in the transition to adding this functionality.

Differential Revision: D31635517

fbshipit-source-id: 4f8d59d1a0235930d8776e5b6bf5b0a4efb3bd04