Continue implementing augmented assignment syntactic rewriter (#1055)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1055

In this diff I continue the work started in the previous; here we implement two more rules for simplifying augmented assignments. They are:

* `complex.attr += any` is rewritten to `a0 = complex; a0.attr += any`
* `a0.attr += any` is rewritten to `a1 = a0.attr ; a1 += any; a0.attr = a1`

In the previous diff we wrote a rule where `a1 += any` is rewritten to `a2 = any; a1 += a2` and so we now have almost all the rules we need to ensure that every augmented assignment only has identifiers on both sides.

We do not yet implement rules for `a[b] += c` and the like; those will come in a later diff.

Reviewed By: wtaha

Differential Revision: D31293128

fbshipit-source-id: 8e911b26607cbe274a2bb4c60918dfdcc236234d