Handle augmented assignment additions in models (#1060)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1060

At long last we correctly extract a graph from models which contain `+=` operators on graph nodes.

This logic is complicated by the fact that `+=` in Python can mean both "mutate the left side" and "add the left side to the right side, assign the result to the left side".

As noted in the comments, it should be rare to have a mutating left side and a graph node right side; we might consider detecting this strange condition and throwing an exception because odds are pretty good that the model is not going to be one we can compile.

If the left side is not mutating then we simply do the addition as normal and return the result.

So far only `+=` is implemented; I'll add `-=` and `*=` and all the rest in an upcoming diff.

Reviewed By: wtaha

Differential Revision: D31382826

fbshipit-source-id: 053658274243a0b5b974088fd3ce179e9926af4f