Test to demonstrate super() compiler bug (#1059)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1059

Models which call super() in a random variable, functional, or callee of a random variable or functional crash; this is due to the way we rewrite function calls.

I've added a test which demonstrates the mechanism of the crash: we handle function calls by intercepting the call, obtaining a reference to the called function, and checking to see if it is one of our known functions. But calls to `super()` cannot be handled this way.

The test also demonstrates the unwanted crash in one of our models. I have not yet come up with a fix for this; a workaround in the meanwhile is to avoid using `super()` and instead do the old fashioned form `super(DerivedClass, self)`

Reviewed By: wtaha

Differential Revision: D31513017

fbshipit-source-id: 01e7b4dd71d04adb641c750bc9eb357998a32822