Implement remaining augmented assignments
Summary:
The beanstalk compiler now handles all augmented assignments, (`+= *= -1 /=`...).

While testing I discovered that unfortunately tensors are mutable in place, which presents a number of problems for a compiler that does not perform alias analysis. For now we will make the assumption that the model does not combine both aliasing and in-place mutation, and therefore we can treat `y += z` the same as `y = y + z` in the compiler, when `y` is tensor-valued or graph-node valued before the mutation.

Differential Revision: D31615528

fbshipit-source-id: 7a4c01020af3a4c294f712391a9e08b698606be6