get_proposers: return a list of proposers from a given world
Summary:
This diff implements the spawning of proposers as previously discussed in [this Quip doc](https://fb.quip.com/WSnUAIDqxHyx#temp:C:UfTd7281592b8e24fd3a6185969d).

Before the change, the `get_proposer` method is only invoked once per chain. However, since the number of nods in `World` can change between iterations, we might need to instantiate new proposers in the middle of an inference. Starting from this diff, `get_proposers` is going to be invoked once per iteration, and the inference classes are responsible for deciding how and when to instantiate a new proposer.

The logic inside `Sampler` is also updated to support running a list of proposers. Since a proposer could be instantiate in the middle of an inference, the remaining number of adaptive iteration is also passed to `get_proposers` so that new proposers can be initialized appropriately.

Not sure if pre-generate a list of proposers and execute each of them in random permutation order is the right way to go, but at least this is consistent with our [single site implementation](https://www.internalfb.com/code/fbsource/[7e3f468344bf4e456869f656e80342747d6bb740]/fbcode/beanmachine/beanmachine/ppl/inference/abstract_mh_infer.py?lines=332-334) :). I feel like the proposer generation logic is probably going to change quite a few times in the future as we figure out what's the best thing to do.

Reviewed By: jpchen, neerajprad

Differential Revision: D31311147

fbshipit-source-id: 9231364252036a3737d5b426f84012bacc2947d1