Implement remaining augmented assignments (#1064)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1064

The beanstalk compiler now handles all augmented assignments, (`+= *= -1 /=`...).

While testing I discovered that unfortunately tensors are mutable in place, which presents a number of problems for a compiler that does not perform alias analysis. For now we will make the assumption that the model does not combine both aliasing and in-place mutation, and therefore we can treat `y += z` the same as `y = y + z` in the compiler, when `y` is tensor-valued or graph-node valued before the mutation.

Reviewed By: wtaha

Differential Revision: D31615528

fbshipit-source-id: 7d82fbdc5a6df0ed1504f45601db46a09ca2570b