Supporting num_chains!=1 across the Python/C++ interface
Summary: In this diff we support passing a parameter of num_chains through BMGInference().infer which is then passed on to the C++ Graph.infer() method. This diff is using a default value of 1 so as to not break existing tests/calls. It would be good to have the same default as Bean Machine's regular inference methods, which is 4. We will make that change in an upcoming diff.

Differential Revision: D31670849

fbshipit-source-id: 524ac2ee3499b78ea5d9f4b73b0bf8382e6cf7b5