Changing default value of num_chains to 4
Summary: With the last few diffs in this stack having introduced num_chains as a parameter to BMGInference().infer, this diff simply changes the default value from 1 to 4, and updates the tests that need to be updated as a result. It should be noted that this change was made only for the .infer method and not the ._infer method, which does not usually (meaning in regular Bean machine methods) take num_chains as a parameter.

Reviewed By: ericlippert

Differential Revision: D31704183

fbshipit-source-id: 6b0c4292d644199dc839fe67b9ca9f7b4a788d0a