Add tests for aliased and undecorated random variables (#1076)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1076

While looking at a bug I noticed a TODO in the code referring to a feature that has already been implemented but was never tested.

Earlier versions of the compiler *required* that a random variable be declared with a specific decorator syntax. We have not required that for some time, but never tested it.

We now have tests which verify:

* If a random variable is declared with a decorator then it can be any alias for `random_variable`; it need not be `bm.random_variable` specifically.
* A random variable can be declared by applying a decorator function directly to an existing function.

Reviewed By: wtaha

Differential Revision: D31803897

fbshipit-source-id: 394b04a196eea7b704987e9998433a4e028f5e1f