Stop rewriting super() calls in single assignment rewriter
Summary:
A call to `super()` is a special syntactic sugar in Python -- when analyzed by the Python runtime it is treated as though it had been written `super(__class__, self)` where `self` is the first parameter to the method.  We need to preserve calls to `super()` in the code and not rewrite them into something like `x = []` / super(*x)` or the like.

In this diff we fix the single assignment rewriter so that it does not rewrite `super()` calls.

Reviewed By: wtaha

Differential Revision: D31808253

fbshipit-source-id: 09b6d9fa9c1b7582fa69d007bc378eee89d14a4f