Making reporting of unattributed time deterministic
Summary:
Steffan noticed that the current implementation of perf reporting adds/drops the field "unattributed" in a manner that depends on total runtime values. This leads to an unexpected leakage of non-determinsim into the output of tests. This diff turns off the dependence of time, and updates existing tests accordingly.

Caveat: Doing this also required eliminating a "-" sign that can sometime appear in santized times by making the unattributed time print an absolute value. This is not entirely ideal for the human reader of the generted text. A TODO item has been added to fix this.

Reviewed By: ericlippert

Differential Revision: D31877189

fbshipit-source-id: 9eb08b0771601d562eeeca77584a9f165b477a15