Making reporting of unattributed time deterministic (#1084)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1084

Steffan noticed that the current implementation of perf reporting adds/drops the field "unattributed" in a manner that depends on total runtime values. This leads to an unexpected leakage of non-determinsim into the output of tests. This diff turns off the dependence of time, and updates existing tests accordingly.

Caveat: Doing this also required eliminating a "-" sign that can sometime appear in santized times by making the unattributed time print an absolute value. This is not entirely ideal for the human reader of the generted text. A TODO item has been added to fix this.

Reviewed By: ericlippert

Differential Revision: D31877189

fbshipit-source-id: 38760c348fcc6f8ce26d5b4aaba7eb91243dc98e