Bind original function's __class__ as outer variable
Summary:
The `super()` function in Python is very special; it is a syntactic sugar for `super(__class__, self)` where `__class__` is an outer variable of the method containing the call to `super()`. Python only automatically generates the closure when a method is known to contain a call to `super()` or use of a variable `__class__`. If you force a call to `super()` using some other syntax, you can trick Python into not generating the closure, and then the call will fail with an obscure `missing __class__ cell` error.

When we transform an original model function into its "lifted" form we close over a variable containing a reference to the "runtime" graph accumulator; we generate code like:

    def foo_helper(bmg):  # bmg refers to the runtime
        def foo(whatever):
            lifted code here can use bmg
        return foo

We then execute foo_helper(bmg) to obtain the inner function bound to the runtime.  But that means that if there was a call to `super()` inside the original function, there is now a call to `super()` inside a completely different function not associated with the original class, and the call will then fail.

What we can do to fix the problem is do the same trick that Python does: **generate an outer variable `__class__` that contains a reference to the original containing class. We already generate an outer helper method; we can simply add a `__class__` parameter to it.

We now do so. See comments in the test case for more details.

Note that this is a special case of a more general problem that we do not yet solve: what happens if an RV (or a callee of an RV) has outer variables?  We lose track of them, and we should not. We'll fix that more general problem in a later diff.

Reviewed By: wtaha

Differential Revision: D31700064

fbshipit-source-id: e2099a10aee5b23de3e929612794ac43e269bc02