Special treatment for Cholesky errors (#1095)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1095

The motivation for this diff is similar to [pyro-ppl/pyro#2009](https://github.com/pyro-ppl/pyro/issues/2009).

Basically when running a GP with MCMC inference it's kind of normal (?) for an invalid proposal to cause a Cholesky decomposition error, in which case we should simply reject the proposal instead of throwing an error and terminate the program.

So in this diff we simply reject a proposal if a cholesky error occur (aka do nothing to update the world). In the future we should probably keep track of the acceptance rate, display it to users in real time, and maybe throw a warning if the acceptance rate is too low.

Reviewed By: neerajprad

Differential Revision: D31916748

fbshipit-source-id: d2cfc12b0af56bc992dc91de95396f1c7bdbe930