Beanstalk now gives error messages for use of @, mm, matmul (#1093)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1093

Though BMG supports a matrix multiplication operator, we have not yet written the code to type check and transform an accumulated graph into a BMG-supported graph. However we are now one step closer: we now actually accumulate the graph in models which use `@` (the Python infix matrix multiplication operator), or either of the torch `mm` or `matmul` functions.

(Note that `mm` and `matmul` have different broadcasting behaviors, so we will likely need to keep track of which was intended in order to generate a correct BMG graph.)

I also changed the output of the DOT generator and the error messages to notate matrix multiplications as `@` rather than `*`.

Reviewed By: wtaha

Differential Revision: D31845800

fbshipit-source-id: 5ceefed994ddcceb749fa89e57c0e07c6e163a56